backend/requirements.txt
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.3
bs4==0.0.2
bsedata==0.6.0
bselib==0.0.5
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.20
django-cors-headers==4.7.0
djangorestframework==3.15.2
exchange_calendars==4.9
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
idna==3.10
korean-lunar-calendar==0.3.1
lxml==5.3.1
nselib==1.6
numpy==2.2.3
outcome==1.3.0.post0
pandas==2.2.3
pandas_market_calendars==4.6.1
playwright==1.50.0
pyee==12.1.1
pyluach==2.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
scipy==1.15.2
selenium==4.29.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
toolz==1.0.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.45.1
wsproto==1.2.0
django_extensions
# Core Data and ML Libraries

numpy
pandas
scikit-learn
xgboost
joblib

# Deep Learning
tensorflow
keras

# Data Retrieval & APIs
yfinance
pandas_datareader
nsetools
requests

# Visualization
matplotlib
streamlit

# LLM and Transformers
google-generativeai
transformers
torch

# Web Scraping (if needed later)
beautifulsoup4
lxml

# Orchestration (if pipelines implemented)
apache-airflow

# Monitoring (optional, for production)
prometheus-client

# For loading .env secrets (optional good practice)
python-dotenv

# Others
tqdm
